+++ test_api.py
@@ -1,8 +1,10 @@
 """ API Tests """
 import unittest
+
+from httmock import HTTMock, all_requests
+
 import woocommerce
 from woocommerce import oauth
-from httmock import all_requests, HTTMock
 
 
 class WooCommerceTestCase(unittest.TestCase):
@@ -14,39 +16,39 @@ class WooCommerceTestCase(unittest.TestCase):
         self.api = woocommerce.API(
             url="http://woo.test",
             consumer_key=self.consumer_key,
-            consumer_secret=self.consumer_secret
+            consumer_secret=self.consumer_secret,
         )
 
     def test_version(self):
-        """ Test default version """
+        """Test default version"""
         api = woocommerce.API(
             url="https://woo.test",
             consumer_key=self.consumer_key,
-            consumer_secret=self.consumer_secret
+            consumer_secret=self.consumer_secret,
         )
 
         self.assertEqual(api.version, "wc/v3")
 
     def test_non_ssl(self):
-        """ Test non-ssl """
+        """Test non-ssl"""
         api = woocommerce.API(
             url="http://woo.test",
             consumer_key=self.consumer_key,
-            consumer_secret=self.consumer_secret
+            consumer_secret=self.consumer_secret,
         )
         self.assertFalse(api.is_ssl)
 
     def test_with_ssl(self):
-        """ Test ssl """
+        """Test ssl"""
         api = woocommerce.API(
             url="https://woo.test",
             consumer_key=self.consumer_key,
-            consumer_secret=self.consumer_secret
+            consumer_secret=self.consumer_secret,
         )
         self.assertTrue(api.is_ssl, True)
 
     def test_with_timeout(self):
-        """ Test timeout """
+        """Test timeout"""
         api = woocommerce.API(
             url="https://woo.test",
             consumer_key=self.consumer_key,
@@ -57,9 +59,8 @@ class WooCommerceTestCase(unittest.TestCase):
 
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            """ URL Mock """
-            return {'status_code': 200,
-                    'content': 'OK'}
+            """URL Mock"""
+            return {"status_code": 200, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
@@ -67,12 +68,12 @@ class WooCommerceTestCase(unittest.TestCase):
         self.assertEqual(status, 200)
 
     def test_get(self):
-        """ Test GET requests """
+        """Test GET requests"""
+
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            """ URL Mock """
-            return {'status_code': 200,
-                    'content': 'OK'}
+            """URL Mock"""
+            return {"status_code": 200, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
@@ -80,24 +81,25 @@ class WooCommerceTestCase(unittest.TestCase):
         self.assertEqual(status, 200)
 
     def test_get_with_parameters(self):
-        """ Test GET requests w/ url params """
+        """Test GET requests w/ url params"""
+
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            return {'status_code': 200,
-                    'content': 'OK'}
+            return {"status_code": 200, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
-            status = self.api.get("products", params={"per_page": 10, "page": 1, "offset": 0}).status_code
+            status = self.api.get(
+                "products", params={"per_page": 10, "page": 1, "offset": 0}
+            ).status_code
             self.assertEqual(status, 200)
 
     def test_get_with_requests_kwargs(self):
-        """ Test GET requests w/ optional requests-module kwargs """
+        """Test GET requests w/ optional requests-module kwargs"""
 
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            return {'status_code': 200,
-                    'content': 'OK'}
+            return {"status_code": 200, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
@@ -105,12 +107,12 @@ class WooCommerceTestCase(unittest.TestCase):
             self.assertEqual(status, 200)
 
     def test_post(self):
-        """ Test POST requests """
+        """Test POST requests"""
+
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            """ URL Mock """
-            return {'status_code': 201,
-                    'content': 'OK'}
+            """URL Mock"""
+            return {"status_code": 201, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
@@ -118,12 +120,12 @@ class WooCommerceTestCase(unittest.TestCase):
         self.assertEqual(status, 201)
 
     def test_put(self):
-        """ Test PUT requests """
+        """Test PUT requests"""
+
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            """ URL Mock """
-            return {'status_code': 200,
-                    'content': 'OK'}
+            """URL Mock"""
+            return {"status_code": 200, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
@@ -131,12 +133,12 @@ class WooCommerceTestCase(unittest.TestCase):
         self.assertEqual(status, 200)
 
     def test_delete(self):
-        """ Test DELETE requests """
+        """Test DELETE requests"""
+
         @all_requests
         def woo_test_mock(*args, **kwargs):
-            """ URL Mock """
-            return {'status_code': 200,
-                    'content': 'OK'}
+            """URL Mock"""
+            return {"status_code": 200, "content": "OK"}
 
         with HTTMock(woo_test_mock):
             # call requests
@@ -144,18 +146,27 @@ class WooCommerceTestCase(unittest.TestCase):
         self.assertEqual(status, 200)
 
     def test_oauth_sorted_params(self):
-        """ Test order of parameters for OAuth signature """
+        """Test order of parameters for OAuth signature"""
+
         def check_sorted(keys, expected):
             params = oauth.OrderedDict()
             for key in keys:
-                params[key] = ''
+                params[key] = ""
 
             ordered = list(oauth.OAuth.sorted_params(params).keys())
             self.assertEqual(ordered, expected)
 
-        check_sorted(['a', 'b'], ['a', 'b'])
-        check_sorted(['b', 'a'], ['a', 'b'])
-        check_sorted(['a', 'b[a]', 'b[b]', 'b[c]', 'c'], ['a', 'b[a]', 'b[b]', 'b[c]', 'c'])
-        check_sorted(['a', 'b[c]', 'b[a]', 'b[b]', 'c'], ['a', 'b[c]', 'b[a]', 'b[b]', 'c'])
-        check_sorted(['d', 'b[c]', 'b[a]', 'b[b]', 'c'], ['b[c]', 'b[a]', 'b[b]', 'c', 'd'])
-        check_sorted(['a1', 'b[c]', 'b[a]', 'b[b]', 'a2'], ['a1', 'a2', 'b[c]', 'b[a]', 'b[b]'])
+        check_sorted(["a", "b"], ["a", "b"])
+        check_sorted(["b", "a"], ["a", "b"])
+        check_sorted(
+            ["a", "b[a]", "b[b]", "b[c]", "c"], ["a", "b[a]", "b[b]", "b[c]", "c"]
+        )
+        check_sorted(
+            ["a", "b[c]", "b[a]", "b[b]", "c"], ["a", "b[c]", "b[a]", "b[b]", "c"]
+        )
+        check_sorted(
+            ["d", "b[c]", "b[a]", "b[b]", "c"], ["b[c]", "b[a]", "b[b]", "c", "d"]
+        )
+        check_sorted(
+            ["a1", "b[c]", "b[a]", "b[b]", "a2"], ["a1", "a2", "b[c]", "b[a]", "b[b]"]
+        )